requirements.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.7.0
billiard==3.6.1.0
black==19.10b0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-debug-toolbar==2.1
django-environ==0.4.5
django-haystack==2.8.1
django-model-utils==4.0.0
django-sphinxdoc==2.0.0
django-taggit==1.2.0
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==3.0.0
filelock==3.0.12
flake8==3.7.9
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
incremental==17.5.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
mccabe==0.6.1
mimesis==3.3.0
more-itertools==8.0.2
oauthlib==3.1.0
packaging==19.2
pathspec==0.6.0
Pillow==6.2.1
pluggy==0.13.1
pockets==0.9.1
priority==1.3.0
py==1.8.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pytest==5.3.2
pytest-benchmark==3.2.2
pytest-django==3.7.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
regex==2019.12.19
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
tox==3.14.2
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.9
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.7.1
celery==4.4.0
redis==3.3.11
flower==0.9.3
django-celery-results==1.1.2
django-celery-email==3.0.0
django-phonenumber-field==4.0.0
django-treebeard==4.3
phonenumbers==8.11.1
psycopg2==2.8.4
